pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "websocket>=0.2.1",
    "websocket-client>=1.8.0",
]
